backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
rasterio==1.3.8
fpdf==1.7.2
firebase-admin==6.2.0
scikit-learn==1.3.2
earthengine-api==0.1.374  # optional for GEE